requirements.txt
spotdl
ttkthemes
pillow
requests
pyinstaller
yt-dlp>=2024.3.10 